refactor(dayjs): Refactor internationalized time formatting feature (#22870) (#22872)
