+++ web/app/components/tools/mcp/provider-card.tsx
@@ -6,7 +6,7 @@ import { useAppContext } from '@/context/app-context'
 import { RiHammerFill } from '@remixicon/react'
 import Indicator from '@/app/components/header/indicator'
 import Icon from '@/app/components/plugins/card/base/card-icon'
-import { useFormatTimeFromNow } from './hooks'
+import { useFormatTimeFromNow } from '@/hooks/use-format-time-from-now'
 import type { ToolWithProvider } from '../../workflow/types'
 import Confirm from '@/app/components/base/confirm'
 import MCPModal from './modal'
